Update README
- Improved project documentation with clearer instructions and setup guidelines.
- Added details on recent features and usage examples.
