+++ Dockerfile
@@ -2,6 +2,8 @@ FROM ultralytics/ultralytics:latest-jetson-jetpack4
 
 WORKDIR /eye-state-detect
 
+RUN pip3 install fastapi uvicorn opencv-python-headless
+
 COPY . .
 
-CMD pip3 list | grep ultralytics
+CMD python3 src/main.py
