+++ requirements.txt
@@ -1,14 +1,20 @@
 absl-py==2.1.0
+annotated-types==0.7.0
+anyio==4.5.2
 astunparse==1.6.3
 attrs==24.2.0
+blinker==1.8.2
 cachetools==5.5.0
 cattrs==24.1.2
 certifi==2024.8.30
 charset-normalizer==3.4.0
+click==8.1.7
 contourpy==1.1.1
 coremltools==8.1
 cycler==0.12.1
 exceptiongroup==1.2.2
+fastapi==0.115.6
+Flask==3.0.3
 flatbuffers==24.3.25
 fonttools==4.55.2
 gast==0.4.0
@@ -16,10 +22,13 @@ google-auth==2.36.0
 google-auth-oauthlib==1.0.0
 google-pasta==0.2.0
 grpcio==1.68.1
+h11==0.14.0
 h5py==3.10.0
 idna==3.10
 importlib_metadata==8.5.0
 importlib_resources==6.4.5
+itsdangerous==2.2.0
+Jinja2==3.1.4
 joblib==1.4.2
 keras==2.13.1
 kiwisolver==1.4.7
@@ -33,6 +42,7 @@ oauthlib==3.2.2
 onnx==1.17.0
 onnxruntime-gpu @ file:///ultralytics/onnxruntime_gpu-1.8.0-cp38-cp38-linux_aarch64.whl#sha256=72d6cfc71ac9b9278ab52fac2b2589c0458f16e58b242cb8b3eeca549dedcad5
 opencv-python==4.10.0.84
+opencv-python-headless==4.10.0.84
 openvino==2024.0.0
 openvino-telemetry==2024.5.0
 opt_einsum==3.4.0
@@ -45,6 +55,8 @@ py-cpuinfo==9.0.0
 pyaml==24.9.0
 pyasn1==0.6.1
 pyasn1_modules==0.4.1
+pydantic==2.10.3
+pydantic_core==2.27.1
 pyparsing==3.1.4
 python-dateutil==2.9.0.post0
 pytz==2024.2
@@ -56,6 +68,8 @@ scikit-learn==1.3.2
 scipy==1.10.1
 seaborn==0.13.2
 six==1.17.0
+sniffio==1.3.1
+starlette==0.41.3
 sympy==1.13.3
 tensorboard==2.13.0
 tensorboard-data-server==0.7.2
@@ -71,11 +85,12 @@ threadpoolctl==3.5.0
 torch @ https://github.com/ultralytics/assets/releases/download/v0.0.0/torch-1.11.0a0+gitbc2c6ed-cp38-cp38-linux_aarch64.whl#sha256=87c3b1fade566123ddb4e1aa3e7a4ad49a5f05c8324556b0d8b40896731bf4ae
 torchvision @ https://github.com/ultralytics/assets/releases/download/v0.0.0/torchvision-0.12.0a0+9b5a3fe-cp38-cp38-linux_aarch64.whl#sha256=71fb38cc7c39e825119f23cd896877b883c5e9fb4a77ca3fabfb35912dd41c60
 tqdm==4.67.1
-typing_extensions==4.5.0
+typing_extensions==4.12.2
 tzdata==2024.2
 -e git+https://github.com/ultralytics/ultralytics@e0f8eda366c7cf08a1c311a17eaaaf87914f8012#egg=ultralytics
 ultralytics-thop==2.0.13
 urllib3==2.2.3
+uvicorn==0.32.1
 Werkzeug==3.0.6
 wrapt==1.17.0
 zipp==3.20.2
